requirements.txt
fastapi
uvicorn
python-multipart
argon2-cffi
sqlalchemy
psycopg2
pydantic-settings
jinja2
